hsd: split up getPublicKey functionality
The getPublicKey method has been split into four separate methods:

- getAccountXpub(int acc, boolean confirm)
- getAddress(int acc, int branch, int index, boolean confirm)
- getPublicKey(int acc, int branch, int index, boolean confirm)
- getXpub(string path)

All the methods save getXpub assume hardened derivation and infer
the coin type from the network flag passed to LedgerHSD.
The getXpub method allows an arbitrary xpub to be retrieved.
A mandatory on-device confirmation will be added to the app for
this method.