all: update library with new parameter logic
GET PUBLIC KEY does not return all xpub details unless specified
in the APDU command parameters. The client api was updated to
accomodate this change.

Also, the parent fingerprint is now properly returned from the
Ledger device. Previously, it was created on the client side
by querying for the parent pubkey and hashing locally.