all: pass options through GET_PUBLIC_KEY code path
All functions involving the GET_PUBLIC_KEY apdu call now take
an optional, options hash. This options hash is passed through
the code logic to the apdu encoding function. This allows for
future options to be added without breaking the API.

Currently, we use this hash to specify which additional details
to return with the public key, i.e. xpub or address. The on-screen
confirmation and network serialization flags are also specified
on this hash.